Add index.html scripts.js and README
Added index.html, scripts.js, and the README file. Created the basic structure of the rock, paper scissors game. Created one function that allows the player to choose, one that allows the computer to choose, and one to start the game, comparing the results and determining a winner or tie.