Add variables and functions to play 5 rounds
Added the pWin and cWin variables to keep score of player and computer wins. Added the game() function to loop through 5 rounds of rock paper scissors by calling the playRound function 5 times and incrementing the pWin or cWin variable based on result. After 5 rounds, the game() function compares pWin to cWin and alerts the result as a win, loss, or tie.